changed ints to floats for all commands and parser functionality

+++ src/commandcontrol/BackCommand.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public class BackCommand extends OneParamCommands {
 	
-	private int myAmount;
+	private double myAmount;
 	
 	public BackCommand(List<Command> l) {
 		super(l);
 	}
 	
-	public int execute(){
+	public double execute(){
 		//move turtle backward
 		return myAmount;
 	}
